Support for 'Services' in KeepAliveRequest
With Every KeepAlive message, a list of currently running services is
sent to the Server. Each Service is identified by (interface, port,
proto, name). Of this `interface` is required for performing 'capture'
later on, `port` and `proto` will be required for adding the `filter`
string during the capture/dissection later on. `name` is for display.